% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.phcsh_mpl.R
\name{fit.phcsh_mpl}
\alias{fit.phcsh_mpl}
\title{Function file for phcsh_mpl}
\usage{
fit.phcsh_mpl(formula, risk, data, control, ...)
}
\description{
This is the function file for model fit for cause-specific hazards model
with maximum penalized likelihood estimation.
}
