% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.phcshcf_mpl.R
\name{fit.phcshcf_mpl}
\alias{fit.phcshcf_mpl}
\title{Function file for phcsh_mpl}
\usage{
fit.phcshcf_mpl(formula, risk, z, data, control, ...)
}
\description{
This is the function file for model fit for cause-specific hazards model
with a cure model using maximum penalized likelihood estimation.
}
