% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phcsh_mpl_control.R
\name{phcsh_mpl_control}
\alias{phcsh_mpl_control}
\title{Ancillary arguments for controling coxph_mpl fits}
\usage{
phcsh_mpl_control(
  knots.perc.limit = c(0.05, 0.95),
  gq.points = 15,
  dgr = 2,
  basis.intercept = FALSE,
  max.outer = 100L,
  max.iter = 10000L,
  lambda,
  n.basis = NULL,
  iter.disp = TRUE,
  iknots.pos = NULL,
  aps = TRUE,
  tmid = TRUE,
  inner.conv = 1e-08
)
}
\arguments{
\item{knots.perc.limit}{test}

\item{gq.points}{test}

\item{dgr}{test}

\item{basis.intercept}{test}

\item{max.outer}{test}

\item{max.iter}{test}

\item{lambda}{test}

\item{n.basis}{test}

\item{iter.disp}{test}

\item{iknots.pos}{test}

\item{aps}{test}

\item{tmid}{test}

\item{inner.conv}{test}
}
\description{
This is used to set various numeric parameters controling a Cox model fit
using phcsh_mpl. Typically it would only be used in a call to phcsh_mpl. Some
basic checks are performed on inputs, such that impossible argument values
(like a negative number of events per base, for example) are avoided.
}
