% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phcsh_mpl_func.R
\name{phcsh_mpl_func}
\alias{phcsh_mpl_func}
\title{Function file for csh mpl}
\usage{
phcsh_mpl_func(formula, risk, data, control, ...)
}
\description{
This is the function file for model fit for cause-specific hazards model
with maximum penalized likelihood estimation.
}
