% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phcshcf_mpl_func.R
\name{phcshcf_mpl_func}
\alias{phcshcf_mpl_func}
\title{Function file for cshcf mpl}
\usage{
phcshcf_mpl_func(formula, risk, z, data, control, ...)
}
\description{
This is the function file for model fit for cause-specific hazards model
with a cure model using maximum penalized likelihood estimation.
}
