% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phcshMPL-package.R
\docType{package}
\name{phcshMPL-package}
\alias{phcshMPL}
\alias{phcshMPL-package}
\title{test me titel}
\description{
Estimate the regression coefficients and the baseline hazard of proportional
hazard Cox models for competing risks data with left, right or interval
censored survival data using maximum penalised likelihood.
A 'non-parametric' smooth estimate of the baseline hazard function is
provided test 2
}
\author{
\strong{Maintainer}: Joseph Descallar \email{jdescallar17@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
