% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.phcsh_mpl.R
\name{plot.phcsh_mpl}
\alias{plot.phcsh_mpl}
\title{Plot a phcsh_mpl object}
\usage{
\method{plot}{phcsh_mpl}(
  object,
  risk = 1,
  plots = c("bh", "surv", "cif"),
  sand = FALSE,
  n.points = 1000,
  pred = NULL
)
}
\arguments{
\item{object}{an object inheriting from class coxph_mpl, representing a
fitted cause specific proportional hazards model.}

\item{risk}{a number which specifies which risk  to plot from phcsh_mpl
object}

\item{plots}{determines which type of plot to produce. bh refers to baseline
hazard plot. surv refers to survival plot, and cif refers to cumulative
incidence function plot. If none are specified, then all three plots are
displayed.}

\item{sand}{use sandwich estimates of standard error}

\item{n.points}{number of points to use for plot. Deafult is n = 1000.}
}
\description{
Plot the bases used to estimate the baseline hazard parameters, as well as
the estimate and confidence interval of the baseline hazard, cumulative
baseline hazard and baseline survival functions (plots are selectable by
which).
}
