% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phcsh_mpl.R
\name{plot_phcsh_mpl}
\alias{plot_phcsh_mpl}
\title{Plot a phcsh_mpl Object}
\usage{
plot_phcsh_mpl(
  object,
  risk = 1,
  plots = c("bh", "surv", "cif"),
  sand = FALSE,
  n.points = 1000
)
}
\arguments{
\item{object}{an object inheriting from class coxph_mpl, representing a
fitted cause specific proportional hazards model.}

\item{risk}{specificy risk}

\item{plots}{which type of plot to produce}

\item{sand}{use sandwich estimates of standard error}

\item{n.points}{number of points to use for plot}
}
\description{
Plot the bases used to estimate the baseline hazard parameters, as well as
the estimate and confidence interval of the baseline hazard, cumulative
baseline hazard and baseline survival functions (plots are selectable by
which).
}
