% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.phcsh_mpl.R
\name{predict.phcsh_mpl}
\alias{predict.phcsh_mpl}
\title{Predict function for a phcsh_mpl object}
\usage{
\method{predict}{phcsh_mpl}(object, covs = NULL, risk = 1, n.points = 1000, sand = FALSE)
}
\arguments{
\item{object}{an object inheriting from class phcsh_mpl, representing a
fitted cause specific proportional hazards model.}

\item{covs}{vector of covariates for prediction of interest. If no
covariates are specificied then the reference values from object are used}

\item{risk}{a number which specifies which risk to plot from phcsh_mpl
object}

\item{n.points}{number of points to use for plot. Deafult is n = 1000.}

\item{sand}{use sandwich estimates of standard error}
}
\description{
Predict the baseline hazards, survival functions and cumulative incidence
functions for a specific covariate
}
