% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.phcsh_mpl.R
\name{summary.phcsh_mpl}
\alias{summary.phcsh_mpl}
\title{Summarise a phcsh_mpl object}
\usage{
summary.phcsh_mpl(object, sand = FALSE, grad = FALSE)
}
\arguments{
\item{object}{An object inheriting from class coxph_mpl, representing a
fitted Cox proportional hazard model.}

\item{sand}{use sandwich estimates of standard error}
}
\description{
Additional information about the Cox proportional hazard model fit
represented by object is extracted and included in the returned object,
which is suitable for printing with the generic print function. The generic
coef function will extract the matrix of coefficients of interest with
standard errors, zz-statistics and p-values.
}
