% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phcshcf_mpl_summary.R
\name{summary_phcshcf_mpl}
\alias{summary_phcshcf_mpl}
\title{Summarise a phcshcf_mpl object}
\usage{
summary_phcshcf_mpl(object, sand = FALSE)
}
\arguments{
\item{object}{An object inheriting from class phcsh_mpl, representing a
fitted proportional hazards cause specific model.}

\item{sand}{use sandwich estimates of standard error}
}
\description{
Additional information about the cause-specifics hazard model fit
represented by object is extracted and included in the returned object,
which is suitable for printing with the generic print function.
}
