% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phcshMPL-package.R
\docType{package}
\name{phcshMPL-package}
\alias{phcshMPL}
\alias{phcshMPL-package}
\title{phcshMPL: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Joseph Descallar \email{jdescallar17@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
